+++ LegendWeathers/Config.cs
@@ -28,9 +28,13 @@ namespace LegendWeathers
         public readonly ConfigEntry<float> bloodMoonAmbientMusicVolume;
         public readonly ConfigEntry<string> bloodMoonAmbientMusicType;
         public readonly ConfigEntry<float> bloodMoonResurrectWaitTime;
+        public readonly ConfigEntry<string> bloodMoonResurrectBlacklistStr;
         public readonly ConfigEntry<bool> bloodMoonStoneItemSpawning;
         public readonly ConfigEntry<string> bloodMoonStoneSpawnBlacklistStr;
+        public readonly ConfigEntry<string> bloodMoonSpecificEnemiesItemSpawningMode;
+        public readonly ConfigEntry<int> bloodMoonSpecificEnemiesItemSpawningChance;
         public readonly ConfigEntry<string> bloodMoonDifficultyFactor;
+        public readonly List<string> bloodMoonResurrectBlacklist = new List<string>();
         public readonly List<string> bloodMoonStoneSpawnBlacklist = new List<string>();
 
         public Config(ConfigFile cfg)
@@ -56,10 +60,13 @@ namespace LegendWeathers
             bloodMoonIntroMusicVolume = cfg.Bind("Blood Moon", "Intro music volume", 0.8f, new ConfigDescription("An introduction music is played when the Blood Moon is spawned on the planet, this will last 30s.\nYou can customize the music volume here.", new AcceptableValueRange<float>(0f, 1f)));
             bloodMoonAmbientMusicVolume = cfg.Bind("Blood Moon", "Ambient music volume", 1f, new ConfigDescription("The volume for the Blood Moon ambient music when you are outside (similar to the vanilla Eclipsed ambient music).", new AcceptableValueRange<float>(0f, 1f)));
             bloodMoonAmbientMusicType = cfg.Bind("Blood Moon", "Ambient music type", "Blood Moon", new ConfigDescription("If you prefer to have the vanilla ambient Eclipsed music instead of the Blood Moon ambient custom music, then you can change it here.\nPlease note that the 'Ambient music volume' config will not apply if you select the Eclipsed music.", new AcceptableValueList<string>("Blood Moon", "Eclipsed")));
-            bloodMoonResurrectWaitTime = cfg.Bind("Blood Moon", "Enemy resurrection wait time", 10f, new ConfigDescription("When an enemy is resurrected by the Blood Moon effect, this is the time in seconds it will take before it actually respawns.", new AcceptableValueRange<float>(0f, 30f)));
+            bloodMoonResurrectWaitTime = cfg.Bind("Blood Moon", "Enemy resurrection wait time", 10f, new ConfigDescription("When an enemy is resurrected by the Blood Moon effect, this is the time in seconds it will take before it actually respawns.\nThe default time value is 10s because it's the expected base difficulty value for the weather.", new AcceptableValueRange<float>(0f, 90f)));
+            bloodMoonResurrectBlacklistStr = cfg.Bind("Blood Moon", "Enemy resurrection blacklist", "", "Comma separated list of enemy names that will never be resurrected by the Blood Moon when they are killed.\nThis MouthDog,Crawler,Flowerman is a valid example, so make sure to only use internal enemies names here.");
             bloodMoonStoneItemSpawning = cfg.Bind("Blood Moon", "Blood Stone spawning", true, "Allow the spawning of the Blood Stone item during specific conditions.");
-            bloodMoonStoneSpawnBlacklistStr = cfg.Bind("Blood Moon", "Blood Stone spawning blacklist", "", "Comma separated list of enemy names that will never spawn Blood Stones when they are killed\nThis MouthDog,Crawler,Flowerman is a valid example, so make sure to only use internal enemies names here.");
-            bloodMoonDifficultyFactor = cfg.Bind("Blood Moon", "Difficulty", "Normal", new ConfigDescription("The difficulty factor of the Blood Moon effect, this will impact the number of spawned enemies during the weather.", new AcceptableValueList<string>("Easy", "Normal", "Hard")));
+            bloodMoonStoneSpawnBlacklistStr = cfg.Bind("Blood Moon", "Blood Stone spawning blacklist", "", "Comma separated list of enemy names that will never spawn Blood Stones when they are killed.\nThis MouthDog,Crawler,Flowerman is a valid example, so make sure to only use internal enemies names here.");
+            bloodMoonSpecificEnemiesItemSpawningMode = cfg.Bind("Blood Moon", "Specific enemies item spawning Mode", "Limited", new ConfigDescription("Choose how enemies that spawn natural items after their death are managed by the weather.\nThis config takes effect on shotguns that are spawned by Nutcrackers and instruments that are spawned by Ghost enemies [from the Haunted_Harpist mod].\n'Limit' will only spawn 1 of these items globally for each apparition of the weather and 'Chance' will get these items a configurable chance value of spawning.", new AcceptableValueList<string>("Limited", "Chance")));
+            bloodMoonSpecificEnemiesItemSpawningChance = cfg.Bind("Blood Moon", "Specific enemies item spawning Chance", 100, new ConfigDescription("If the 'Specific enemies item spawning Mode' config is set to 'Chance', this config will define the percentage chance of spawning the specified items when the enemy dies.\nA value of 50 means there is a 50% chance of spawning the item.", new AcceptableValueRange<int>(0, 100)));
+            bloodMoonDifficultyFactor = cfg.Bind("Blood Moon", "Difficulty", "Normal", new ConfigDescription("The difficulty factor of the Blood Moon effect, this will impact the number of spawned enemies during the weather.", new AcceptableValueList<string>("Easy", "Normal", "Hard", "Extreme")));
 
             cfg.Save();
             cfg.SaveOnConfigSet = true;
@@ -90,11 +97,19 @@ namespace LegendWeathers
 
         private void PopulateList()
         {
-            if (string.IsNullOrWhiteSpace(bloodMoonStoneSpawnBlacklistStr.Value))
-                return;
-            foreach (string value in bloodMoonStoneSpawnBlacklistStr.Value.Split(',').Select(s => s.Trim()))
+            if (!string.IsNullOrWhiteSpace(bloodMoonStoneSpawnBlacklistStr.Value))
             {
-                bloodMoonStoneSpawnBlacklist.Add(value.ToLower());
+                foreach (string value in bloodMoonStoneSpawnBlacklistStr.Value.Split(',').Select(s => s.Trim()))
+                {
+                    bloodMoonStoneSpawnBlacklist.Add(value.ToLower());
+                }
+            }
+            if (!string.IsNullOrWhiteSpace(bloodMoonResurrectBlacklistStr.Value))
+            {
+                foreach (string value in bloodMoonResurrectBlacklistStr.Value.Split(',').Select(s => s.Trim()))
+                {
+                    bloodMoonResurrectBlacklist.Add(value.ToLower());
+                }
             }
         }
     }